override plugin version in the provider interface (#19043)
For Git plugins we *always* want to use the version from the Git repo,
be that a tag or a commit hash. We currently override it for the purpose
of SDK generation in project_sdk_link.go, however we don't override it
in GetPluginInfo. This means the user gets a warning that their plugin
version isn't correct.

To fix this consistently, override the version in the provider plugin,
overriding the version in the schema itself, and also in GetPluginInfo
if we have a Git Plugin.

---------

Co-authored-by: Fraser Waters <fraser@pulumi.com>