Update Go program gen to target 1.23 (#19026)
We use program gen to generate programs for users as part of convert. We
generally want these programs to be targeting recent versions of the
language so that users can start building off them quickly with modern
features.
For example all our dotnet programs target dotnet 8.
Go was targeting 1.20 but there are a number of new features and
language changes in later version of Go that we'd want users to be able
to use easily, so this bumps program gen to target 1.23 (the older of
the two Go versions we currently support).