add dependencies field to the schema
Add a dependencies field that can be used to set dependencies across
languages.  Currently we only support doing language specific
dependencies, however at least for components and conformance tests it
would be nicer to not have to specify the dependencies per language,
but instead specify them once and let codegen do the rest.

This PR implements that for `pulumi` dependencies only (though the
schema would allow for namespaces to be added).

This implements that in the schema and adds it to TypeScript codegen.
Other languages to follow.