l3-component-simple test (#17832)
Adds a simple conformance test that component resources work. We added
support for these for converting TF modules, but this is also good for
testing that the runtimes treat component resources the same way.

Currently this only tests Typescript because I ran into programgen
issues for both Go and Python.

Go is missing a load of types to go from literals to Input types. 
Python is outputting TypedDicts with `total=false` that then don't type
check when accessed.