Add a toOutput intrinsic to pcl and a tree rewriter to utilize it.
Currently ForExpressions as a part of ResourceComponents do not generate
correctly in typescript (and likely other languages).

This is because--at least in the typescript case--a pulumi.Input<T> can
be a T, pulumi.Output<T>, or Promise<T>.  toOutput forces the value to
always be treated as an output, since there is already a method in the
sdk to always convert any value to an output, and the generated code
will always work as expected (whether the input was an Output from
another part of the program or just a normal variable).

This is accomplished with 3 parts:
1. Add a toOutput pcl intrinsic which will use the language SDK to
   convert to an output.
1. When generating a for expression, if it is an output, generate the
   appropriate apply calls.
1. When generating a component be sure to to tree rewriting to call
   toOutput on all ForExpression variables which come from config
   variables of the Component.

Fixes #19060