fix names in type references when namespaces are defined in package.json (#19066)
When we introduced namespaces in TypeScript, we missed that the Analyzer
was still using `packageJSON.name` to generate type references. This
meant types would be generated with just the name, but references would
be broken, because they would use the full name field from
`package.json`, which can include the namespace.

Fix that by passing the name through correctly.

Fixes https://github.com/pulumi/pulumi/issues/19055