Speed up workspace.DetectProjectFromPath
There is a utility function workspace.DetectProjectFromPath that attempts to scan the file system starting at a given
path to find the deepest containing ancestor directory that appears to have a Pulumi.yaml or a similar project file
present.

Before this change it did not scale well for situations where ancestor folders have a lot of entries. It would scan
each ancestor directory and examine every entry to test it for being the Pulumi project file.

For example, in a `pulumi convert` scenario a Mac M3 Pro was taking 10-15 seconds to perform 300000 file system checks
here in a situation where tempdirs had a lot of entries, dominating the cost of any other calls.

With this change the function no longer scans directories but instead probes for possible Pulumi.yaml file locations
more directly.