Improve watch output (#19083)
This change improves the watch command's output to make it a little more
engaging, informative, and useful. Namely:

- Emit the fully qualified stack name.

- Print the update permalink in the usual "View Live" manner.

- Use a consistent color scheme with SpecHeadline, with some specific
uses of colors, bolds, and underlines, to make things visually pop.

- Eliminate stray newlines that can lead to double newlines depending on
the outputs of the program, language host, or providers.

- Colorize the watch events resource names gutter, to make it easier to
see at a glance the interleaved message streams associated with certain
resources. This is just done by hashing the name and indexing into a
list of possible colors to ensure it is deterministic.

- Improve the "starting" and "done" event formatting so it's easier to
eyeball and correlate the two. Also distinguish between success and
failure for every resource operation.

- Print the stack outputs after an update completes.

- Display a summary of what operations, and how many, were performed.

Here is an example before:


![image](https://github.com/user-attachments/assets/cdc53988-860d-4176-9cff-a6abe39da55d)

And after:


![image](https://github.com/user-attachments/assets/2bcdce61-0fa6-41e3-9f33-545f6ad89d32)

---------

Co-authored-by: Justin Van Patten <jvp@justinvp.com>
Co-authored-by: Fraser Waters <fraser@pulumi.com>