Add Refresh program support to the engine (#18868)
This adds two new features to the step generator, which together allow
you to run a `pulumi refresh` operation (if it was actually plumbed into
the CLI layer) by running the user program.

The first feature is the ability to run refreshes on resources before
processing the rest of their step generation logic. This could be a
replacement for `preview/up --refresh` but currently is only used for
the full program refresh mode. This works similarly to the parallel diff
logic where we send a step (a real step in this case because we do want
the refresh reported and snapshot'd) and then when it completes it sends
an event back to the step generator.

The second feature is the ability to skip all other custom resource
operations, so we _only_ refresh, this also triggers refresh operations
for anything left in state that was registered at the end of the
deployment, as opposed to triggering deletes.

---------

Co-authored-by: Will Jones <will@sacharissa.co.uk>