Add l1-output-null test (#19001)
`null` is a valid value in the PCL model, and stacks should be able to
return outputs of null directly, or with null nested inside other
structures. Unfortunately that breaks a load of the rest of the model
where we treat optional resource args which in the runtime language will
just be null as not sent fields. So for now this test just checks that
we can write null in PCL and return an array of null.

This also changes the l1-builtin-try test to return an array of null,
rather than null directly to work around the above issue.

---------

Co-authored-by: Brandon Pollack <brandonpollack23@gmail.com>