speed up python performance tests with this one simple trick (#19099)
In the python conformance tests we build wheels for the Python SDK, and
the SDK for every provider. We use `python build` for this. `python
build` builds a sdist and then a wheel from that by default. However we
don't need the source distribution at all in conformance tests and only
care about the wheel. Instructing `python build` to only build the wheel
cuts the build time roughly in half. This brings the time to run a
single python conformance test locally from 48 seconds to 27 seconds.

---------

Co-authored-by: Julien <julien@caffeine.lu>