exclude virtualenv from being checked by mypy (#19107)
Mypy doesn't exclude the virtualenv by default when typechecking. It
does exclude only hidden files, starting with a `.`. At the same time,
`uv` installs a file into the virtualenv that doesn't typecheck
correctly.

By default that's fine, as `uv` uses `.venv` as its virtualenv location,
but it breaks when someone uses a custom venv location, and uses mypy.
For example in conformance tests, we use `venv` as the virtual env
location, which in turn makes the tests fail because of `mypy`. This
could also happen to someone using a custom venv location with `uv`.