cache SDK artifacts
In conformance tests we curently generate all the SDKs once, but we
re-pack them for each test run.  This seems unnecessary, as the
packed artifacts shouldn't change.  Only pack them once, to save time.
Also `uv` seems to be non-deterministic when packing these artifacts,
which seems to result in flakes in https://github.com/pulumi/pulumi/pull/19109.

Cache the artifacts instead, and re-use them for various tests.

Note that we currently simply overwrite artifacts in the same
directory, so if the artifacts were any different, our tests would
already be broken, since any test we're running would get a random
artifact to use.