install dependencies in python conformance tests using uv (#19109)
In previous PRs we started using `uv` more in conformance tests, e.g.
for building the Python wheel, which got us to run the conformance tests
in 24 minutes in CI. That still seemed kinda slow, so I did some more
digging, and found that we are still installing all our dependencies
using `pip`.

This PR changes the python conformance tests, so they install
dependencies via `uv` instead. This assumes `uv` is available, but I
expect that to be the case in developer environments.

This brings the total time for the CI job running the conformance tests
to ~16min, shaving off another 8 minutes.