Make `property.Value` immutable (#17500)
tldr: making all `property.Value`s immutable will make the API less
idiomatic, but it will make API design and usage of `property.Value`
clearer and less bug prone. We can do it without sacrificing too much
performance.

This change is stacked on top of
https://github.com/pulumi/pulumi/pull/17374.

There are two major concerns to address with this PR:

- The API is unidiomatic. ☝️ explains why I think the change is worth it
from that perspective.
- The change is not performant.

## API Usage

Regardless of this PR, all APIs that manipulate `property.Value` *must*
be of the shape:

```go
func alterValue(startingValue property.Value) (mutatedValue property.Value, _ error)
```

This is because you cannot reliably mutate `startingValue`. You can only
mutate `startingValue` if it is a map or an array.[^1] APIs of this
shape essentially must discard `startingValue` after `alterValue` is
called, since they cannot be sure what value it holds, or what relation
it has with `mutatedValue`.

Other languages can express this as an inout (C++) value, or as passing
an owned value in and requiring an owned value out (Rust). In Go, we
could express this with an in-out `*property.Value`, but that is highly
unidiomatic:

```go
func alterValue(value *property.Value) error
```

By making `property.Value` immutable, we side-step this entire
conversation. The idiomatic Go function is correct:

```go
func alterValue(property.Value) (property.Value, _ error)
```

The calling API is now free to reuse the value it passed in. The
immutable nature of the API makes APIs like `func(property.Value)`
clearly wrong, as opposed to flakey.

## Performance

As long as we keep all values immutable, we only need shallow copies on
`.AsMap/.AsArray`, and *never* need to make deep copies.
`property.Value.Equals` can depend on internals to make target deep
mutations as efficient as possible. A full manual traversal of a nested
datastructure has the same cost as copying.

I believe that we can afford the additional cost of copying to prevent
bugs. TF's underlying data value is immutable (implemented the same way)
and TF does not suffer on performance. IMO, the immutability of
`cty.Value` makes it much easier to work with.

[^1]: If you're sure that the func will only accept an array or a map,
it should encode that in the type system.