Properties used in component outputs are never plain
The plain attribute is only relevant for inputs. When generating the
schema for a component, we need to track wether we are currently looking
at a type used as an input or an output of the component. The properties
of types used as component inputs, including nested types, as well as
list items and dictionary values, can potentially be plain, unless they
are within an input. The properties of types used as component outputs
can never be plain.

There is some churn in the tests because we were previously sharing
complex types for both inputs and outputs, but that's not possible due
to the restriction on the plain property.

Fixes https://github.com/pulumi/pulumi/issues/18694