Pick the right `refresh` when running programs (#19140)
This might be the result of a botched rebase or the like, but we're
currently picking the wrong refresh method (and consequently iteration
source) when passing `--run-program`, which causes a crash if the source
gets touched (since the default is an error source). This change fixes
that so that `--run-program` works in all cases for `refresh`.