Wait longer for sentinel file in TestRunCanceled
This test continues to be flaky, however I’ve been unable to reproduce this locally. Presumably it flakes under heavy load/resource contention.

When I looked into this previously, I concluded (possibly wrongly) that we get `signal: interrupt` when we cancel the context before the operation has started. If we cancel after the operation has started we get the expected exit status.

The previous attempt in https://github.com/pulumi/pulumi/pull/18178/files to unflake this test used a sentinel file that we look for so that we know we’ve started the operation. This waited for up to 10 seconds for this file. This PR increases this to a full minute.

We can the flake rate and most recent occurence  here https://app.codecov.io/gh/pulumi/pulumi/tests/master?term=TestRunCanceled