partition the tests module by test instead of by package (#19137)
The tests module takes a long time to run, so let's allow it being split
by test instead of running the whole package at once.

This leaves the NodeJS automation API tests as the slowest ones. We also
reduce the number of tests run at the same time from 8 partitions for
`tests` (that was overkill though) to 6 (4 for smoke tests + 2 for
everything else that's in `tests`)