Format cmd/pulumi-language-python-exec
This file is gitignored at the top level (apparently for goreleaser?),
but changes are tracked in git because the file was intentionally added
to git at some point. However tools like ruff don't understand that
distinction and ignore the file (because of the .gitignore). Un-ignore
the file at the sdk/python level, add it to the ruff includes and format
it.