don't override schema version in `package *`
Since https://github.com/pulumi/pulumi/pull/19043 we're overriding
this when getting the schema, so we don't need to override it here.

In fact overriding it here causes issues e.g. for terraform providers,
which use the version in the schema differently.

```
$ pulumi package get-schema terraform-provider hashicorp/random 3.6.0 | jq -r '.version'
3.6.0

$ pulumi package get-schema terraform-provider@0.10.0 hashicorp/random 3.6.0 | jq -r '.version'
0.10.0
```

The second call here should also 3.6.0, not 0.10.0.  This PR makes it so.