Avoid creating a main.py file when chosing the uv toolchain (#19154)
We run `uv init` to create the `pyproject.toml` file when using the uv
as the toolchain. Before version 0.6 this would always create a
`hello.py` file, which we then delete. With version 0.6 this changed to
`main.py`, but this also introduced a new `--bare` option to avoid
creating such a file altogether.

Fixes https://github.com/pulumi/pulumi/issues/19153