Fix dependency graphing in the presence of deleted resources
When deleting resources as part of an operation, we must start at the
ends of the dependency graph and proceed inwards. That is, if a resource
`B` depends on a resource `A`, we must delete `B` *first*, lest we have
an intermediate state where there is a dangling reference from a
remaining `B` to a now-deleted `A`. Computing these deletion orders is
done by building a `DependencyGraph` and calling its
`TransitiveDependenciesOf` method. Under the hood,
`TransitiveDependenciesOf` works by building a lookup table that maps
from URNs in a snapshot to resource states, before walking the snapshot
to build a set of dependencies. This works fine, until we encounter a
snapshot where URNs are (contrary to their name) *not* enough to
uniquely identify resources!

Consider the following snapshot, with simplified URNs for readability:

```
[
  {
    "urn": "A"
  },
  {
    "urn": "B",
    "dependencies": ["A"]
  },
  {
    "urn": "A",
    "delete": true
  }
]
```

The URN `A` appears twice, but this is allowed -- the second has
`"delete": true`, which indicates it is a resource which has been
deleted but not cleaned up from the snapshot. This can happen during
e.g. a create-before-replace operation, where a new `A` is created
before the old one is deleted. In such operations, the old `A` is marked
`"delete": true` so that the new one can be added, before being cleaned
up later on. If that clean-up fails, the presence of the deleted
resource means that Pulumi can sort it as part of a future (hopefully
successful) operation. To re-iterate -- this is fine, and snapshot
integrity checks permit any number of deleted copies of a resource.

Unfortunately, `TransitiveDependenciesOf` does not budget for this
possibilty, and prior to this change will build an *incorrect* set of
dependencies in which `B` above would be reported as depending on the
*deleted* `A`. This plays havoc with deletions later on, especially in
the presence of `--continue-on-error` or `protect`, which permit an
operation to continue despite failures early on. When this happens, we
need to cancel deletions on resources which are depended on by things
that we failed to delete. This only works if our dependency graphs are
correct -- if not, we end up permitting deletions on e.g. the new `A`
above, meaning we end up with a snapshot where `B` has a dangling
reference.

This change fixes `TransitiveDependenciesOf` to track old (`"delete":
true`) and new resources separately, as we do elsewhere in the codebase
(e.g. when topologically sorting snapshots). This is not perfect since
as mentioned, there could be multiple deleted copies of a resource, but
in the absence of actually unique identifiers for resources in a
snapshot, it's the best we can do. This fix addresses what we suspect is
the cause of snapshot integrity issues when using `--continue-on-error`.

> [!TIP]
>
> Found with fuzzing

Fixes #19114