Save package source to `packages` in Pulumi.yaml on `package add` (#19157)
This PR adds a side effect to the `pulumi package add` command. It will
now modify (or add) the `packages` section in Pulumi.yaml to point to
the added package. It will fill the following properties:

- `source` will be set to whatever the first parameter of `package add`
command was, but without a version
- `version` will be set to the schema version or parameterized
provider's base provider version
- `parameters` will be set to the parameter values passed to `package
add`

The command will use either the short or the long version of the
reference with some heuristics to pick the right one:

- Short by default
- Long if parameters were set
- Long if there was a reference to the same package already and it had
long form
- Long if there were other references and they were all long

### Key Implementation Details

1. The InstallPackage function now returns the schema package in
addition to any error
2. The package_add command has been updated to:

  - Extract source and version information from the package
  - Build a PackageSpec with appropriate details
  - Add the package to the project via a new AddPackage method
  - Save the updated project back to Pulumi.yaml

3. A new AddPackage method was added to the Project type which:

  - Determines whether to use string format or PackageSpec format
  - Preserves existing format when updating an existing package
  - Ensures consistency in format usage across multiple packages

4. Tests have been added/updated to verify:

- Different package addition scenarios (simple, versioned,
parameterized)
  - Format consistency maintenance
  - Package replacement behavior
  - Format conversion when parameters are added

Resolves https://github.com/pulumi/pulumi/issues/19142 but without the
`--save` option. Users can call `pulumi package gen-sdk` if they want to
generate an SDK with no other side effects.

I believe codecov's number is underreported since a bunch of testing
happens in integration.