fix package_add panic when there's no version in the schema
In https://github.com/pulumi/pulumi/pull/19157 we introduced saving
details of the package added with `pulumi package add` into the
Pulumi.yaml packages section.  For doing that we get the version from
the installed package, which in turn gets it from the schema.

However the version is an optional field in the schema.  We set it
automatically when using Git based plugins, but e.g. local plugins
don't necessarily return a version at all.  This means `pkg.Version`
can be `nil` and thus the CLI panics.  Add a check for that, and keep
the version from what's passed in in the command line args if available.