Replace ParseRange with ParseVersion (#19149)
ParseRange doesn't handle pre-release and build metadata parts properly
and we're not actually using ranges, just version equality checks.

This allowed us to change the conformance test to add build metadata to
the provider version, this revealed an issue with Go where it doesn't
like the build metadata in the go.mod so we clean it up when writing
out.