stack_test: make each test copy testproj to its own temp dir
Multiple tests are using this test project for their testing, and drop
temporary config files there in some cases.  This leads to flaky
tests, such as in https://github.com/pulumi/pulumi/issues/19194.

Fix that by copying the test project to a temporary directory for each
of the tests, which makes sure they do not interfere with eachother.