Add bootstrap-less mode for Python component providers
We now support two ways of running Python based plugins: bare
directories (the existing implementation) and now additionally as
buildable packages.

If the plugin directory is a bare directory (that is not a Python
package) with a `__main__.py`, we run the plugin's `__main__.py`
directly. Dependencies for these plugins are specified via
`requirements.txt`.

Otherwise, we check if the plugin directory is a buildable Python
package. In that case we run the plugin via `pulumu.run.plugin`.
Dependencies are installed from the `dependencies` section of the
`pyproject.toml`.