Fix ReplaceOnChanges and IgnoreChanges being respected during DBR diffs (#19188)
Fixes #19056

This fixes two bugs with DBR logic. Firstly that ReplaceOnChanges would
never trigger a replace as part of DBR even if the field was being
changed, secondly that IgnoreChanges wouldn't prevent a replace if a
field changed (but shouldn't have because of ignore changes).

Doing this requires that ReplaceOnChanges is saved to state with
IgnoreChanges so that we have access to it at the point DBR is checking
dependencies from state to see if they need replacing.