fix coverage for integration tests (#18719)
We're currently not calculating coverage from integration tests. This is
because we set both the `GOCOVERDIR` environment variable, and use the
`-cover` option in `go test`. Unfortunately `-cover` overrides
`GOCOVERDIR` to something Go sets internally, with no way to override
that using flags or environment variables. See also
https://github.com/golang/go/issues/66225.

We can work around that by setting a `PULUMI_GOCOVERDIR` variable
instead, and turning that into `GOCOVERDIR` whenever we execute
`pulumi`. This is somewhat awkward, but I think it's the best we can do
while this upstream bug is not fixed.

Fixes https://github.com/pulumi/pulumi/issues/18331