Tidy up python component provider tests
Previously the tests installed the dependencies for the component
provider inside the test data folder. This was done because ProgramTest
updates the provider section in Pulumi.yaml to an absolute path to the
original provider.

Now that we have a packages section in Pulumi.yaml, which remains
relative, the tests can copy the provider code into the temporary test
directory, and we can run the dependency installation for each test.

This gives us better test isolation.