register provider for external component resources
Component resources can be created in two ways, either in the users
program directly, or in a provider running in a separate process.

Currently we don't track the provider relationship for the second case
at all.  However in reality these components are created by a
provider, and we should track that relationship. The component
resource itself doesn't know whether it is being created in a program,
or externally in a provider.

However if it is being created externally, we know we already
registered a provider for the package in the engine, for the Construct
request.  We can make use of this fact, and register that provider for
the component resource if it exists.  Since there is not going to be a
provider for component resources that are created in a program, they
will continue behaving as before.