Add run-program to automation api (#19218)
Adds --run-program support to nodejs, python, and go.

I've added this explicitly (that is we explicitly pass `=true` or
`=false`) so that if we change the default behaviour of the flag in the
future anyone who explicitly set `false` will continue to get the same
behaviour they expect. As opposed to passing `--run-program` for true
and nothing for false, where if we then changed the default behaviour
anyone who had explicitly set false would suddenly be having their
program run.