Use PackageSpec.Dependencies for component provider dependencies (#19236)
With https://github.com/pulumi/pulumi/pull/19071 we can pass
`dependencies` to the schema and let codegen handle language specific
details, instead of doing so inside the component provider host and
abusing language specific deps.

This also fixes an issue where the dependencies from the schema were not
being set when using `setup.py`, as well as an inconsistency on
`typing-extensions`.