Don't discard non-outputty component outputs (#19228)
Instead of expecting a component’s outputs to be isntances of `Output`,
we check the shcema, and pick those properties.

Additionally, this fixes an where `plain` was being set on component
outputs. The `plain` property should never be set on any type that’s
used as a component output. See also
https://github.com/pulumi/pulumi/pull/19135.

Fixes https://github.com/pulumi/pulumi/issues/18960