Handle local plugin paths with a version number
In https://github.com/pulumi/pulumi/pull/18940 we extended `pulumi install` to generate the SDKs for packages referenced in `Pulumi.yaml`, however depending on how the provider is referenced in `Pulumi.yaml`, this does not work for local providers.

When a provider is added via `pulumi package add /some/path`, we create an entry in `Pulumi.yaml` with the version:

```
packages:
  greeter: /some/path@0.0.0
```

This causes `pulumi install` to fail, as it does not properly detect the path due to the version number.

Fixes https://github.com/pulumi/pulumi/issues/19253