Add resourceType/Name to TS (#18842)
TypeScript parts for https://github.com/pulumi/pulumi/issues/15890

We can't copy the other languages and add a new method/property on the
Resource class itself because of SxS issues. Users may have multiple
copies of the pulumi SDK due to old provider SDKs. As such anything new
on resource has to be typed optional, but the ergonomics of using that
are awkward.

Instead this adds two new top level functions `resourceName` and
`resourceType` that are both `Resource -> string`. If a resource from an
old SDK is passed to `resourceName` it will throw that `__name` is
undefined. The type function is ok because I think we've had
`__pulumiType` since forever.