Don’t install python/nodejs runtimes during `make build` (#19250)
Similar to https://github.com/pulumi/pulumi/pull/19048 we should not
install the binaries during `make build`.

This helps when using `mise`. Previously `make build` would also install
`pulumi-language-python` into `$GOBIN`, however not the
`pulumi-language-python-exec` shim, causing problems when the shim is
not found in the same directory as the runtime.