Implement `fmt.GoStringer` for `property.{Value,Map,Array}` (#19259)
Stacked on #19258.

This PR implements `fmt.GoStringer` for `property` types. This enables
programmers to paste the value from `fmt.Printf("Properties: %#v",
myPropertyMap)` into a unit test and have it compile. This is a quality
of life feature for Go programmers using the library. It should not
effect the runtime behavior of `property`.