sdk/{nodejs,python}/Makefile: export GOBIN before using it (#19278)
In https://github.com/pulumi/pulumi/pull/19250 we stopped setting
`GOBIN` for the `dist` target in the makefiles for
`sdk/{nodejs,python}`. This is fine for the `go install` part, but are
also using the same env variable to drop more scripts into this
directory. Having it not set breaks this target. Fix that by re-adding
setting the `GOBIN` variable for the target.