fix: detect readme in subdir for package publishing (#19298)
When publishing a VCS based component - via `pulumi package publish` -
it picked the README in the root of the repository instead of the one
located in the subdirectory where the component is located.

For example, running `GITHUB_TOKEN="$(gh auth token)" pulumi package
publish
https://github.com/pulumi/pulumiup-2025-keynote-demo/components/container-app
--source private` should use this readme:
https://github.com/pulumi/pulumiup-2025-keynote-demo/blob/main/components/container-app/README.md

But it ends up picking this one:
https://github.com/pulumi/pulumiup-2025-keynote-demo/blob/main/README.md

The missing piece was taking the packages `LocalName` into account as
well.

Fixes https://github.com/pulumi/pulumi-service/issues/27663

---------

Co-authored-by: Thomas Gummerer <t.gummerer@gmail.com>