remove "Debug branch name" step (#19308)
This shouldn't be needed anymore (I think we introduced this
temporarily, to debug selectively upload test results to codecov).
Somehow this failed in
https://github.com/pulumi/pulumi/actions/runs/14643049132/job/41090458864?pr=19305.

It really doesn't make sense to me that it failed (Windows behaves
weirdly sometimes I guess), but at the same time we don't need this
step, so let's get rid of it.