Add pulumi.type_token decorator (#19309)
Add a `pulumi_type` decorator to annotate Pulumi types such as resources
and enums with their type token. This also provides a `get_pulumi_type`
helper to retrieve the token.

Once this is released, we can update codegen to use this decorator to
annotate resources and enums. We will also update the codegen change
introduced in https://github.com/pulumi/pulumi/pull/19221 to use the
decorator instead.

Ref https://github.com/pulumi/pulumi/issues/19297 &
https://github.com/pulumi/pulumi/issues/18484