Ensure we always save ImportID (#19314)
`ImportID` was only set when doing an import which meant it was lost
from state if you did a normal update on the resource. This then broke
the re-import logic on a later `up`.

This fixes it to always save it to state, and updates the test to check
that actual import IDs and IDs are kept track of ok.