Use the new property.Value system for plugin config (#19285)
This updates the language and analyser plugin interfaces to take
`property.Map` from the engine layer, rather than `map[string]string`.

We still have to translate to string maps internally for backwards
compatibility, but we can push that lower and lower till we can
eventually get rid of it.

There's a lot of follow ups that could be done after this to spread the
new `property.Value/Map` system further, but that's not a rush. We can
translate between the two at each boundary point for now.