Remove the 'deletes' field, just use 'dones' (#19284)
Rather than tracking deletes as a special case we can just treat them
the same as things being "done" generally. This is really useful when we
start allowing multiple operations per-resource (e.g. import then
update) as the done semantics just work.

There were comments about "If this is a delete-replace operation, we
don't want to mark the resource as deleted because we want to keep the
new resource." but I can't recall the reasoning for those comments and
removing them hasn't broken any tests. If anyone can come up with a test
case that depended on these things share, but otherwise I think this is
safe.

The only other change needed was when we make `SameSteps` the two state
pointers have to be different objects. This requires a copy in
`getDependencySteps`. I've also added an assert to the same step
constructor to ensure we don't break that.