Pass kind to RunPlugin (#19320)
We'll be making use of this for both the analyzer work we're doing but
also it's useful the debugging code proposed for attaching to
components.

In general languages like TypeScript and Python need to know what sort
of plugin they're running because they have bootstrap scripts that they
run before the user code itself, e.g. "@pulumi/pulumi/cmd/run-plugin"
for provider plugins.