treat private plugins as Git Plugins still
In IsLocalPluginPath, we check if a plugin is a remote plugin. To do
that, we parse the PluginSpec first, and then check based on that.

However, when the plugin is in a private Git repo, we get an error
here, because we fail to fetch the latest tag of the Git repo.
Instead, return a special error for that, and look for it specifically
here, so we can correctly determine that it's a local plugin, or
remote, and thus give a good error message to the user down the line.