Add Handshake to analyzer plugins (#19328)
This adds the handshake method to analyzer plugins. This has a number of
advantages for passing data back and forth.

Firstly it expose a channel for analysers to report things back to the
engine, we don't currently make use of that but we _could_ similar to
provider plugins where they report if they understand new property types
(e.g. if we added sets or integers).

Secondly it allows the engine to attach to a running analyzer, again
we're not currnetly making use of that, but if the issue comes up it's
now easy to solve in the same way we do for providers.

Thirdly it's a well typed and documented list of things that are sent to
analyzer plugins rather than dealing with undocumented environment
variables set in engine code.

Fourthly it passes config as typed property maps rather than
`map[string]string` allowing us to cleanly pass things like `true` vs
`"true"`.

Fifthly it allows us to pass down directory information. Current
analyzer plugins always have to run in the analyzer directory rather
than the pulumi programs directory, which _every_ other plugin runs in.
Passing the directory information explicitly like this means new
analyzer plugins will be able to start in the pulumi program directory
but still find their directory if needed.

Sixthly this unblocks binary policy packs and using RunPlugin for policy
packs. We're not concerned about doing the former yet, but the latter is
a priority before doing dotnet/java/go support.