Update some import tests (#19340)
In preparation for the new import system this changes a few tests such
that they are still valid today, but have extra checks that we want to
assert on for the new system as well.

`TestContinueOnErrorImport` is changed to explicitly fail the read,
rather than relying on the diff failure that is being removed in the new
system.

`TestImportOption` adds some checks that the outputs we see in the
program are what we expect to see.