parse error at plugin startup as gRPC error (#19312)
When we start up a plugin and either it just fails to start up
completely, or we fail reading from it, we're printing the error we got.
Unfortunately if that's a gRPC error, the output is a bit ugly, as it
always includes `rpc error: code = Unknown desc = <description>`.

As an example, when `pulumi package add` fails to start up the plugin,
the error is

`error: failed to get schema: could not read plugin
[../broken-package/pulumi-resource-broken-package] stdout: rpc error:
code = Unknown desc = problem executing plugin program (could not run
language executor): fork/exec
/home/tgummerer/test/broken-package/venv/bin/python: no such file or
directory`.

We can easily do at least a little better here, and parse the gRPC
error, which then allows us to print it in a slightly better way,
removing the unhelpful gRPC details:

`error: failed to get schema: could not read plugin
[../broken-package/pulumi-resource-broken-package]: problem executing
plugin program (could not run language executor): fork/exec
/home/tgummerer/test/broken-package/venv/bin/python: no such file or
directory`