Always save fully qualified stack names to settings (#19336)
Noticed in
https://github.com/pulumi/pulumi/pull/19324/files/aa1c465e08639c1578cc43e62fa7ea86879a8634#r2059990017

We need to save fully qualified stack names to the settings to prevent
the following issue.

Imagine the default org is currently "foo" and we create and save a
stack "bar". Historically this would save the abbreviated name to the
settings so just "bar".
If we then change the default org to "qux" and run `up` it will load the
name from settings and reconstruct "bar" into "qux/bar" instead of
"foo/bar" which is the stack that had actually been selected!

This fixes that to just always save the fully qualified name to stack
settings.