factor the option arguments to `mod.typeString`
We pass a few different bools into this function. Let's make that a
struct instead for the options, which is easier to expand in the
future (I'm planning to add another option for https://github.com/pulumi/pulumi/issues/19272),
and has the additonal benefit that it's more self documenting, rather
than needing comments for every argument.