Use server cache value
We store 2 additional pieces of information in the versions cache file: the duration we should keep version information cached, and the timestamp of the last time we prompted the user.

To determine when we should query for a new version, we read the cache duration from the file, and add it to the modification date of the file. If that time is in the past, we query.

To determine if enough time has elapsed to prompt the user, we look at the last prompt timestamp, and see if it is 24 hours in the past, or 1 hour in the past for dev versions.